Cleanup of fix of issue where size of original image determined size of zoom dialog
